# src/main/java/com/management/candidate/domain/LoginResponse.java
package com.management.candidate.domain;

import lombok.Data;

@Data
public class LoginResponse {

    private String token;

    public LoginResponse(String token) {
        this.token = token;
    }
}